requirements.txt
# Core
python-dotenv>=1.0.0
requests>=2.31.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0

# UI
streamlit>=1.31.0

# LangChain ecosystem (v1 compatible)
langchain>=0.1.16
langgraph>=0.0.40
langchain-groq>=0.1.3

# Data / validation
pydantic>=2.6.0
